rename defaults to styled
